server/requirements.txt
argcomplete
flask
flask-socketio
munch
PyYAML
pyOpenSSL
